Add files via upload
This file uses the libaries
- requests
- bs4 import BeautifulSoup 
- csv
This script uses requests to request data from a web browser (URL). The html is then returned using r.text. The html of the page is the parsed using BeautifulSoup. The script then wbescrapes all the free proxies within the table which is then written into a csv file